pyproject.toml
[project]
name = "rl-market-making"
version = "0.1.0"
description = ""
authors = [
    {name = "flxclxc",email = "pisacane.alexei@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.7.1,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "hyperopt (>=0.2.7,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
